project/pyproject.toml
[tool.poetry]
name = "project"
version = "0.1.0"
description = ""
authors = ["Tobi-De <tobidegnon@protonmail.com>"]
license = "MIT"

[tool.poetry.dependencies]
python = "^3.10"
fastapi = "^0.75.2"
uvicorn = "^0.17.6"
pydantic = {extras = ["email", "dotenv"], version = "^1.9.0"}
fastapi-users = "^10.0.2"
arq = "^0.22"
tortoise-orm = {extras = ["asyncpg"], version = "^0.19.0"}
aerich = "^0.6.3"
sentry-sdk = "^1.5.11"
rich = "^12.4.0"
typer = "^0.4.1"
fastapi-pagination = {extras = ["tortoise"], version = "^0.9.3"}
watchgod = "^0.8.2"

[tool.poetry.dev-dependencies]
pytest = "^7.1.2"
factory-boy = "^3.2.1"
pre-commit = "^2.19.0"
mypy = "^0.950"
isort = "^5.10.1"
flake8 = "^4.0.1"
poetryup = "^0.6.2"
ipython = "^8.3.0"

[tool.aerich]
tortoise_orm = "app.db.config.TORTOISE_ORM"
location = "./migrations"
src_folder = "./."

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
